Refactor. Create package for editor, and its subpackages. Consolidate rows into their own functions.
